+++ ChangeLog.md
@@ -13,6 +13,7 @@ and this project adheres to
 
 ### Added
 
+- function `swap-vals!` for atoms 
 - the option to pass arguments for opening MacOS apps. E.g.: 
   (shell/open-macos-app "TextEdit" "example.txt")
 - an ESR module for dealing with Swiss ESR reference numbers